# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REALIZED API",
    "version": "1.0.0",
    "summary": "What a liquidity provider actually made, not what the DEX advertised.",
    "description": "Advertised LP APR is fee income annualized. It has no price term, so it is positive by construction and cannot report a loss. REALIZED adds the missing term: realized return = fees + impermanent loss, computed from The Graph's historical poolDayData.\n\nNo API key is required. This server proxies its own Graph key, so an agent can call every endpoint below with a bare GET. CORS is open (`access-control-allow-origin: *`).\n\nHonesty constraints that hold across every endpoint:\n- Unmeasurable is reported as `measurable: false` with a reason, NEVER as 0. A fabricated zero is a number whose label lies, which is the exact defect this project exists to expose.\n- Corpus aggregates carry a stable-pair canary: stable/stable pairs must show ~0 impermanent loss, or the instrument is considered broken and refuses to rank.\n- `/api/rank` is a TRACK RECORD, not a forecast. Predictive pool selection was tested and failed walk-forward validation five times against a pre-registered bar; it was removed rather than shipped.",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://realized.drainfun.xyz", "description": "Live production" }
  ],
...
  "paths": {
...
    "/api/find": {
      "get": {
        "tags": ["discovery"],
        "operationId": "findPool",
        "summary": "Resolve a pair name to a poolId",
        "description": "Searches the cached corpus first, then falls back to a live Uniswap v3 mainnet lookup unless `live=0`. Pass `matches[].poolId` into /api/pool/{poolId} or /api/position/{poolId}.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" }, "example": "WETH/USDC", "description": "Pair query, e.g. WETH/USDC or a single symbol." },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 5 } },
          { "name": "live", "in": "query", "required": false, "schema": { "type": "string", "enum": ["0"] }, "description": "Set to 0 to suppress the live fallback and search the cache only." }
        ],
        "responses": {
          "200": {
            "description": "Matches, with the source that produced them",
            "content": { "application/json": { "schema": {
              "type": "object",
              "properties": {
                "query": { "type": "string" },
                "source": { "type": "string", "description": "`cache` or `live:uniswap-v3/mainnet`.", "examples": ["cache", "live:uniswap-v3/mainnet"] },
                "matches": { "type": "array", "items": { "$ref": "#/components/schemas/Match" } },
                "note": { "type": "string" }
              }
            } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/pool/{poolId}": {
      "get": {
        "tags": ["measurement"],
        "operationId": "getPool",
        "summary": "Realized vs advertised return for one pool over a trailing window",
        "description": "Computes realized return (fees + impermanent loss) for a pool over the last `days`, at an assumed range width. For a specific entry date rather than a trailing window, use /api/position/{poolId}.",
        "parameters": [
          { "$ref": "#/components/parameters/PoolId" },
          { "name": "range", "in": "query", "required": false, "schema": { "type": "number" }, "description": "Assumed range width multiplier. 1.25 = tight, 2 = moderate, 4 = wide, 1e8 ≈ full-range (v2-equivalent). Omit for the server default.", "example": 2 },
          { "name": "days", "in": "query", "required": false, "schema": { "type": "integer", "default": 30 } },
          { "$ref": "#/components/parameters/Dex" },
          { "$ref": "#/components/parameters/Chain" }
        ],
        "responses": {
          "200": { "description": "Realized measurement for the pool", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PoolMeasurement" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/wallet/{address}": {
      "get": {
        "tags": ["discovery"],
        "operationId": "getWalletPositions",
        "summary": "Which pools an address holds, and at what REAL tick range",
        "description": "Position DISCOVERY, not per-position P&L — and that distinction is a measured decision, not a missing feature.\n\nReads the Position entity and resolves each position's real tick bounds, so you can stop assuming ±2x. Feed `positions[].range.widthX` into /api/position/{poolId} to get realized return at the range the LP actually set.\n\nWhat it deliberately does NOT return, and why:\n- **Uncollected fees.** `collectedFees*` is a withdrawal record, not an earnings record — it only populates when the LP calls collect(). Measured live: of 150 positions with `collectedFeesToken0 == 0`, 71 had non-zero `feeGrowthInside0LastX128` — they earned fees and never collected. So uncollected fees return `measurable: false` with a reason, never $0.\n- **Position value at entry.** `deposited*`/`withdrawn*` are lifetime cumulative, not entry state (75% of live positions read as one-sided because of it), so they are namespaced under `cumulative` and must not be read as position value.\n\nEvery response carries a `limits` block stating both. Token `decimals` ship beside every raw amount so no consumer has to guess the divisor. Read-only: no signing, no wallet connection, no permissions.",
        "parameters": [
          { "name": "address", "in": "path", "required": true, "schema": { "type": "string", "pattern": "^0x[a-fA-F0-9]{40}$" }, "description": "Owner address. Read-only lookup.", "example": "0x5777d92f208679db4b9778590fa3cab3ac9e2168" },
          { "$ref": "#/components/parameters/Dex" },
          { "$ref": "#/components/parameters/Chain" },
          { "name": "includeClosed", "in": "query", "required": false, "schema": { "type": "boolean", "default": false }, "description": "Include positions with zero current liquidity." }
        ],
        "responses": {
          "200": { "description": "Positions held, with real ranges and explicit limits", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WalletPositions" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/audit": {
      "get": {
        "tags": ["corpus"],
        "operationId": "auditPools",
        "summary": "Corpus-wide sweep with a stable-pair canary and a sensitivity table",
        "description": "Sweeps the corpus and reports what share of pools advertised a positive APR while LPs realized a negative return. Ships a sensitivity table across three liveness gates, because a headline that only survives one cutoff is a parameter, not a finding.",
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 250 } },
          { "name": "days", "in": "query", "required": false, "schema": { "type": "integer", "default": 30 } },
          { "$ref": "#/components/parameters/Dex" },
          { "$ref": "#/components/parameters/Chain" }
        ],
        "responses": {
          "200": { "description": "Audit summary with canary and sensitivity", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Audit" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/rank": {
      "get": {
        "tags": ["corpus"],
        "operationId": "rankPools",
        "summary": "Which advertised APRs have actually been trustworthy (track record, NOT a forecast)",
        "description": "Ranks pools by track record over the measured window. This is explicitly not a prediction: predictive pool selection was tested (+1.64pp on one split) and then failed five walk-forward re-runs against a pre-registered bar of ≥0.5pp median edge, ≥60% positive windows, p ≤ 0.05. It was killed rather than shipped. The response carries `notAPrediction` so an agent cannot mistake it.",
        "parameters": [
          { "name": "top", "in": "query", "required": false, "schema": { "type": "integer", "default": 25, "maximum": 200 } },
          { "name": "trust", "in": "query", "required": false, "schema": { "type": "string", "enum": ["historically honest", "gap-prone", "routinely misleading", "unmeasurable"] }, "description": "Filter to a single trust label." },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string", "enum": ["realized", "gap", "tvl"], "default": "realized" } }
        ],
        "responses": {
          "200": { "description": "Ranked pools with trust-label definitions and counts", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Rank" } } } }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "PoolId": { "name": "poolId", "in": "path", "required": true, "schema": { "type": "string" }, "description": "Pool contract address. Get one from /api/find.", "example": "0x88e6a0c2ddd26feeb64f039a2c41296fcb3f5640" },
      "Dex": { "name": "dex", "in": "query", "required": false, "schema": { "type": "string", "enum": ["uniswap-v3", "aerodrome"], "default": "uniswap-v3" } },
      "Chain": { "name": "chain", "in": "query", "required": false, "schema": { "type": "string", "enum": ["mainnet", "arbitrum", "polygon", "optimism", "base"], "default": "mainnet" } }
    },
    "responses": {
      "BadRequest": { "description": "Missing or invalid parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Pool not found at this dex/chain", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
...
      "Match": {
        "type": "object",
        "properties": { "pair": { "type": "string" }, "poolId": { "type": "string" }, "tvl": { "type": "number" }, "fee": { "type": "number" } }
      },
...
      "PoolMeasurement": {
        "type": "object",
        "properties": {
          "pair": { "type": "string" },
          "poolId": { "type": "string" },
          "days": { "type": "integer" },
          "advertisedAprPct": { "type": "number" },
          "feeReturnPct": { "type": "number" },
          "impermanentLossPct": { "type": "number" },
          "realizedReturnPct": { "type": "number" },
          "realizedAprPct": { "type": ["number", "null"] },
          "measurable": { "type": "boolean" }
        }
      },
...
      "WalletPositions": {
        "type": "object",
        "properties": {
          "owner": { "type": "string" },
          "dex": { "type": "string" },
          "chain": { "type": "string" },
          "openPositions": { "type": "integer" },
          "pools": { "type": "array", "items": { "type": "object", "additionalProperties": true }, "description": "Distinct pools this address holds." },
          "positions": { "type": "array", "items": { "$ref": "#/components/schemas/WalletPosition" } },
          "feeReporting": { "type": "object", "properties": {
            "measurable": { "type": "integer" },
            "earnedButUncollected": { "type": "integer", "description": "Positions that provably earned fees and never collected them. These are reported as not-measurable, never as $0." },
            "note": { "type": "string" }
          } },
          "limits": { "type": "object", "description": "Ships in EVERY response so the caller cannot miss it.", "properties": {
            "measured": { "type": "string" },
            "weReport": { "type": "array", "items": { "type": "string" } },
            "weDoNotReport": { "type": "array", "items": { "type": "string" } },
            "why": { "type": "string" }
          } },
          "next": { "type": "string", "description": "The suggested follow-up call, with the real range width substituted in." }
        }
      },
...
    }
  }
}